requirements.txt
APScheduler==3.6.3
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
python-telegram-bot==13.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.1
wrapt==1.12.1
